// src/print/operations/item/printOperationItemRequestBuilderDeleteRequestConfiguration.ts
import { type RequestOption } from '@microsoft/kiota-abstractions';

export interface PrintOperationItemRequestBuilderDeleteRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}